+++ src/main/java/com/springboot/MyTodoList/repository/CategoryRepository.java
@@ -15,5 +15,5 @@ import javax.transaction.Transactional;
 @Transactional
 @EnableTransactionManagement
 public interface CategoryRepository extends JpaRepository<Category,Integer> {
-    List<Category> findByProjectId(Long projectId);
+    List<Category> findByProjectId(int projectId);
 }
