+++ src/main/java/com/springboot/MyTodoList/service/CategoryService.java
@@ -28,7 +28,7 @@ public class CategoryService {
         }
     }
 
-    public ResponseEntity<List<Category>> getCategoriesByProject(Long projectId) {
+    public ResponseEntity<List<Category>> getCategoriesByProject(int projectId) {
         List<Category> shortcuts = categoryRepository.findByProjectId(projectId);
         if (shortcuts.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
